refactor C modules
